third_party/WebKit/LayoutTests/platform/win/fast/canvas/canvas-blending-text-expected.txt

Series of tests to ensure correct results on applying different blend modes when drawing text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing blend mode source-over
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
FAIL actualColor(4, 4)[1] should be within 5 of 129. Was 255.
PASS actualColor(4, 4)[2] is within 5 of 129
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode multiply
PASS actualColor(4, 4)[0] is within 5 of 129
FAIL actualColor(4, 4)[1] should be within 5 of 129. Was 255.
FAIL actualColor(4, 4)[2] should be within 5 of 65.25882352941176. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode screen
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
PASS actualColor(4, 4)[1] is within 5 of 255
FAIL actualColor(4, 4)[2] should be within 5 of 192.74117647058821. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode overlay
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
PASS actualColor(4, 4)[1] is within 5 of 255
PASS actualColor(4, 4)[2] is within 5 of 130.48235294117646
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode darken
PASS actualColor(4, 4)[0] is within 5 of 129
FAIL actualColor(4, 4)[1] should be within 5 of 129. Was 255.
PASS actualColor(4, 4)[2] is within 5 of 129
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode lighten
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
PASS actualColor(4, 4)[1] is within 5 of 255
PASS actualColor(4, 4)[2] is within 5 of 129
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode color-dodge
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
PASS actualColor(4, 4)[1] is within 5 of 255
FAIL actualColor(4, 4)[2] should be within 5 of 255. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode color-burn
PASS actualColor(4, 4)[0] is within 5 of 129
PASS actualColor(4, 4)[1] is within 5 of 255
FAIL actualColor(4, 4)[2] should be within 5 of 5.930232558139517. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode hard-light
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
PASS actualColor(4, 4)[1] is within 5 of 255
PASS actualColor(4, 4)[2] is within 5 of 130.48235294117646
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode soft-light
FAIL actualColor(4, 4)[0] should be within 5 of 181.3697880023021. Was 129.
PASS actualColor(4, 4)[1] is within 5 of 255
PASS actualColor(4, 4)[2] is within 5 of 129.61611515296823
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode difference
PASS actualColor(4, 4)[0] is within 5 of 126
FAIL actualColor(4, 4)[1] should be within 5 of 126. Was 255.
FAIL actualColor(4, 4)[2] should be within 5 of 0. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode exclusion
PASS actualColor(4, 4)[0] is within 5 of 126
FAIL actualColor(4, 4)[1] should be within 5 of 126. Was 255.
PASS actualColor(4, 4)[2] is within 5 of 127.48235294117649
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode hue
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
FAIL actualColor(4, 4)[1] should be within 5 of 181. Was 255.
FAIL actualColor(4, 4)[2] should be within 5 of 181. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode saturation
PASS actualColor(4, 4)[0] is within 5 of 129
PASS actualColor(4, 4)[1] is within 5 of 255
PASS actualColor(4, 4)[2] is within 5 of 129
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode color
FAIL actualColor(4, 4)[0] should be within 5 of 255. Was 129.
FAIL actualColor(4, 4)[1] should be within 5 of 181. Was 255.
FAIL actualColor(4, 4)[2] should be within 5 of 181. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

Testing blend mode luminosity
FAIL actualColor(4, 4)[0] should be within 5 of 92. Was 129.
FAIL actualColor(4, 4)[1] should be within 5 of 218. Was 255.
FAIL actualColor(4, 4)[2] should be within 5 of 92. Was 129.
PASS actualColor(4, 4)[3] is within 5 of 255

PASS successfullyParsed is true

TEST COMPLETE
